packages/rag/pyproject.toml
[project]
name = "regknots-rag"
version = "0.1.0"
description = "RegKnots RAG chat engine"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.0.0",
    "asyncpg>=0.30.0",
    "tiktoken>=0.7.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
